+++ R/tableby.internal.R
@@ -55,11 +55,13 @@ format.tbstat <- function(x, digits = NULL, ...)
 }
 
 #' @export
-format.tbstat_countpct <- function(x, digits.count = NULL, digits.pct = NULL, ...)
+format.tbstat_countpct <- function(x, digits.count = NULL, digits.pct = NULL, digits = NULL, ...)
 {
   att <- attributes(x)
-  if(is.null(att$which.pct)) att$which.pct <- 0
-  x <- vapply(seq_along(x), function(i) formatC(x[i], digits = if(i %in% att$which.pct) digits.pct else digits.count, format = "f"), NA_character_)
+  x <- vapply(seq_along(x), function(i) {
+    d <- if(i %in% att$which.pct) digits.pct else if(i %in% att$which.count) digits.count else digits
+    formatC(x[i], digits = d, format = "f")
+  }, NA_character_)
   x <- trimws(x)
   attributes(x) <- att
   NextMethod("format")
@@ -73,13 +75,18 @@ format.tbstat_countpct <- function(x, digits.count = NULL, digits.pct = NULL, ..
 #' @param sep The separator between \code{x[1]} and the rest of the vector.
 #' @param parens A length-2 vector denoting parentheses to use around \code{x[2]} and \code{x[3]}.
 #' @param sep2 The separator between \code{x[2]} and \code{x[3]}.
-#' @param pct The symbol to use after percents.
+#' @param pct For statistics of length 2, the symbol to use after the second one. (It's called
+#'   "pct" because usually the first statistic is never a percent, but the second often is.)
 #' @param which.pct Which statistics are percents? The default is 0, indicating that none are.
+#' @param which.count Which statistics are counts? The default is everything except the things that are percents.
 #' @param ... arguments to pass to \code{as.tbstat}.
 #' @details
+#'   The vignette has an example on how to use these.
+#'
 #'   \code{as.tbstat} defines a tableby statistic with its appropriate formatting.
 #'
-#'   \code{as.countpct} adds another class to \code{as.tbstat} to use different "digits" arguments. See \code{\link{tableby.control}}.
+#'   \code{as.countpct} adds another class to \code{as.tbstat} to use different "digits" arguments
+#'   (i.e., \code{digits.count} or \code{digits.pct}). See \code{\link{tableby.control}}.
 #'
 #'   \code{as.tbstat_multirow} marks an object (usually a list) for multiple-row printing.
 #' @name tableby.stats.internal
@@ -96,9 +103,9 @@ as.tbstat <- function(x, oldClass = NULL, sep = NULL, parens = NULL, sep2 = NULL
 
 #' @rdname tableby.stats.internal
 #' @export
-as.countpct <- function(x, ..., which.pct = 0L)
+as.countpct <- function(x, ..., which.count = setdiff(seq_along(x), which.pct), which.pct = 0L)
 {
-  tmp <- as.tbstat(x, ..., which.pct = which.pct)
+  tmp <- as.tbstat(x, ..., which.count = which.count, which.pct = which.pct)
   class(tmp) <- c("tbstat_countpct", class(tmp))
   tmp
 }